Refactor InfoService
The code that constructs the Info object in InfoService has
been moved into PKIEngine such that it can be reused.
